requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.10,<3.0
fhir.resources>=7.1,<8.0
pandas>=2.2,<3.0
orjson>=3.10,<4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
recordlinkage==0.16
presidio-analyzer>=2.2,<3.0
presidio-anonymizer>=2.2,<3.0
spacy>=3.7,<4.0
openai>=1.51,<2.0
dbt-core>=1.8,<2.0
dbt-postgres>=1.8,<2.0
pytest>=8.3,<9.0
httpx>=0.28,<1.0
